requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
Flask==3.1.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.2.0
pillow==11.2.0
propcache==0.3.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
qrcode==8.0
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.1
Werkzeug==3.1.3
yarl==1.18.3
